requirements.txt
streamlit==1.28.0
groq==0.4.1
PyPDF2==3.0.1
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0